--- armi/testing/buildEmptyOperators.py
@@ -1,104 +1,0 @@
-# Copyright 2026 TerraPower, LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""
-This module contains herlper functiosn to build Operators of filled with emtpy Hex or Cartesian blocks. These should
-only be used in testing.
-"""
-
-from armi import operators, settings, tests
-from armi.materials import uZr
-from armi.reactor import assemblies, blocks, grids
-from armi.reactor.components import Hexagon, Rectangle
-
-
-def buildOperatorOfEmptyHexBlocks(customSettings=None):
-    """
-    Builds a operator w/ a reactor object with some hex assemblies and blocks, but all are empty.
-
-    Doesn't depend on inputs and loads quickly.
-
-    Parameters
-    ----------
-    customSettings : dict
-        Dictionary of off-default settings to update
-    """
-    cs = settings.Settings()  # fetch new
-    if customSettings is None:
-        customSettings = {}
-
-    customSettings["db"] = False  # stop use of database
-    cs = cs.modified(newSettings=customSettings)
-
-    r = tests.getEmptyHexReactor()
-    r.core.setOptionsFromCs(cs)
-    o = operators.Operator(cs)
-    o.initializeInterfaces(r)
-
-    a = assemblies.HexAssembly("fuel")
-    a.spatialGrid = grids.AxialGrid.fromNCells(1)
-    b = blocks.HexBlock("TestBlock")
-    b.setType("fuel")
-    dims = {"Tinput": 600, "Thot": 600, "op": 16.0, "ip": 1, "mult": 1}
-    c = Hexagon("fuel", uZr.UZr(), **dims)
-    b.add(c)
-    a.add(b)
-    a.spatialLocator = r.core.spatialGrid[1, 0, 0]
-    o.r.core.add(a)
-    o.r.sort()
-    return o
-
-
-def buildOperatorOfEmptyCartesianBlocks(customSettings=None):
-    """
-    Builds a operator w/ a reactor object with some Cartesian assemblies and blocks, but all are empty.
-
-    Doesn't depend on inputs and loads quickly.
-
-    Parameters
-    ----------
-    customSettings : dict
-        Off-default settings to update
-    """
-    cs = settings.Settings()  # fetch new
-    if customSettings is None:
-        customSettings = {}
-
-    customSettings["db"] = False  # stop use of database
-    cs = cs.modified(newSettings=customSettings)
-
-    r = tests.getEmptyCartesianReactor()
-    r.core.setOptionsFromCs(cs)
-    o = operators.Operator(cs)
-    o.initializeInterfaces(r)
-
-    a = assemblies.CartesianAssembly("fuel")
-    a.spatialGrid = grids.AxialGrid.fromNCells(1)
-    b = blocks.CartesianBlock("TestBlock")
-    b.setType("fuel")
-    dims = {
-        "Tinput": 600,
-        "Thot": 600,
-        "widthOuter": 16.0,
-        "lengthOuter": 10.0,
-        "widthInner": 1,
-        "lengthInner": 1,
-        "mult": 1,
-    }
-    c = Rectangle("fuel", uZr.UZr(), **dims)
-    b.add(c)
-    a.add(b)
-    a.spatialLocator = r.core.spatialGrid[1, 0, 0]
-    o.r.core.add(a)
-    o.r.sort()
-    return o
